fix: refactor spotify router
