+++ fastapi/app/routers/__init__.py
@@ -2,8 +2,7 @@ from fastapi import APIRouter
 
 from .health import health_router
 from .message import message_router
-from .project import project_router
-from .spotify import spotify_music_router
+from .project import project_router, spotify_music_router
 
 api_router = APIRouter()
 api_router.include_router(project_router, tags=["project"], prefix="/projects")
